+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# 1.3.4
+- Added compat with knightlib 2.0.0
+- Attempt to fix a case where the mutated/holy teddy and the golden allay would phase through the ground on certain cases
+
 # 1.3.3
 - Minimum required version of knightlib increased to 1.6.2
 - Fixed demon flesh appearing in every entity loot table in JER
-- 
+
 # 1.3.2
 - Fixed companion attack logic receiving infinite cooldowns, which prevented several companions from dealing damage and Cornelius from completing its taming behavior
 - Fixed hostile imps not pursuing their current targets, as well as Dinamo hitting duplicate targets